Fixing wrong extraction flow
If the entry to extract represents a directory, the check for being a
directory and the creation of the directory is connected within the some
if-statement, which causes subsequent actions to be executed, which are
only necessary, if the entry is a file.
This fix separates the check and the creation such that only the
necessary actions for a directory will be executed, of the entry
represents a directory.