Addition of custom pre-commit hook generation
The repository class now adds a custom Git pre-commit hook with desired
actions (defined by parameter) to the extracted repository. This
addition required writing files, which the additional file utilities
provide (including the reading of files, which is required later
anyway).
The addition of a pre-commit hook only works for Git repositories,
currently, and may be extended to support SVN repository as well later.