Split of repository setup and hook addition
In order to enable the execution of evaluations with the same repository
(instance), but with different configurations of the tool or action
called as part of a repository hook, the addition of that should not be
part of the initial repository setup.
Now, that addition is possible via a new public method, which also
requires the ability to throw execution exceptions instead of setup
exceptions as this method is not related to the core setup anymore.